a suite with composite type, why key encoder spec overwritten

+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/state/StateStore.scala
@@ -298,8 +298,8 @@ object KeyStateEncoderSpec {
           asInstanceOf[List[_]].map(_.asInstanceOf[Int])
         RangeKeyScanStateEncoderSpec(keySchema, orderingOrdinals)
       case "PrefixKeyScanStateEncoderSpec" =>
-        val numColsPrefixKey = m("numColsPrefixKey").asInstanceOf[Int]
-        PrefixKeyScanStateEncoderSpec(keySchema, numColsPrefixKey)
+        val numColsPrefixKey = m("numColsPrefixKey").asInstanceOf[BigInt]
+        PrefixKeyScanStateEncoderSpec(keySchema, numColsPrefixKey.toInt)
     }
   }
 }
